backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flasgger==0.9.7.1
python-dotenv==1.0.0
boto3==1.28.0
botocore==1.31.0